Adopt ISC license
Versions up to and including 2.0.0 were published under the WTFPL,
which caused a few emails from conscientious but confused developers
wanting to make sure that their intended use of the library met my
personal interpretation of "whatever the fuck you want".

The ISC license appears more common and better understood, so
versions subsequent to 2.0.0 will be published under that instead.

Under my interpretation, the only legal effect of this change is
that the license doc must be retained when copying this repo.

(Meh.)